feat: user list and role policy

+++ app/Http/Controllers/Admin/RoleController.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\Admin\UpsertRoleRequest;
 use Inertia\Inertia;
-
 use Spatie\Permission\Models\Role;
 
 class RoleController extends Controller
@@ -12,6 +12,7 @@ class RoleController extends Controller
 
 	public function index()
 	{
+		Gate::authorize('viewAny', Role::class);
 		return Inertia::render('admin/Roles', [
 			'pagination' => Inertia::always(Inertia::merge(Role::orderBy(request('sortField', 'created_at'), request('sortDirection', 'desc'))->paginate(request('perPage', 5), "*", null, request('page', 1)))),
 		]);
@@ -20,13 +21,15 @@ class RoleController extends Controller
 
 	public function store(UpsertRoleRequest $request)
 	{
-		Role::create($request->all());
+		Gate::authorize('store', Role::class);
 
+		Role::create($request->all());
 		return redirect('/admin/roles');
 	}
 
 	public function update(UpsertRoleRequest $request, string $id)
 	{
+		Gate::authorize('update', Role::class);
 		$data = Role::findOrFail($id);
 		$data->update($request->all());
 
@@ -35,6 +38,7 @@ class RoleController extends Controller
 
 	public function destroy(string $id)
 	{
+		Gate::authorize('delete', Role::class);
 		$row = Role::findOrFail($id);
 		$row->delete();
